+++ src/main/java/net/frozenblock/lib/registry/FrozenLibRegistries.java
@@ -36,6 +36,7 @@ import net.frozenblock.lib.levelgen.material.impl.RuleSourceAddition;
 import net.frozenblock.lib.music.api.structure.StructureMusic;
 import net.frozenblock.lib.sound.api.predicate.SoundPredicate;
 import net.frozenblock.lib.sound.api.type.MovingSoundType;
+import net.frozenblock.lib.sound.impl.damage.PlayerDamageTypeSound;
 import net.frozenblock.lib.wind.disturbance.WindDisturbanceType;
 import net.frozenblock.lib.wind.extension.WindManagerExtensionType;
 import net.minecraft.core.HolderLookup;
@@ -86,6 +87,7 @@ public class FrozenLibRegistries {
 	public static final ResourceKey<Registry<ClipGroup>> CLIP_GROUP = ResourceKey.createRegistryKey(FrozenLibConstants.id("clip_group"));
 	public static final ResourceKey<Registry<WaterLikeType>> WATER_LIKE_TYPE = ResourceKey.createRegistryKey(FrozenLibConstants.id("water_like_type"));
 	public static final ResourceKey<Registry<RuleSourceAddition>> RULE_SOURCE_ADDITION = ResourceKey.createRegistryKey(FrozenLibConstants.id("rule_source_addition"));
+	public static final ResourceKey<Registry<PlayerDamageTypeSound>> PLAYER_DAMAGE_TYPE_SOUND = ResourceKey.createRegistryKey(FrozenLibConstants.id("player_damage_type_sound"));
 
     public static void init() {
 		DynamicRegistries.registerSynced(SOUND_TYPE_OVERRIDE, SoundTypeOverride.DIRECT_CODEC);
@@ -96,6 +98,7 @@ public class FrozenLibRegistries {
 		DynamicRegistries.registerSynced(CLIP_GROUP, ClipGroup.DIRECT_CODEC);
 		DynamicRegistries.registerSynced(WATER_LIKE_TYPE, WaterLikeType.DIRECT_CODEC);
 		DynamicRegistries.register(RULE_SOURCE_ADDITION, RuleSourceAddition.DIRECT_CODEC);
+		DynamicRegistries.registerSynced(PLAYER_DAMAGE_TYPE_SOUND, PlayerDamageTypeSound.DIRECT_CODEC);
     }
 
 	public static HolderLookup.Provider vanillaRegistries() {